+++ Mods ChoGGi/OnMsg Print/metadata.lua
@@ -25,5 +25,6 @@ UIPropertyChanged
 See also:
 https://github.com/ChoGGi/SurvivingMars_CheatMods/blob/master/Tutorials/OnMsgs%20Load%20Game.md#load-game
 https://github.com/ChoGGi/SurvivingMars_CheatMods/blob/master/Tutorials/OnMsgs%20New%20Game.md#new-game
+https://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/OnMsg%20Print
 ]],
 })